config/requirements.txt
# ── Data ingestion ──────────────────────────────────────
polygon-api-client>=1.13.0
yfinance>=0.2.38
websockets>=12.0
aiohttp>=3.9.0
praw>=7.7.1
finnhub-python>=2.4.19
requests>=2.31.0
httpx>=0.27.0
alpaca
alpaca-py

# ── Computation ──────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0
pandas-ta>=0.3.14b
scipy>=1.12.0
statsmodels>=0.14.1
pyarrow>=15.0.0
redis>=5.0.0

# ── AI / ML ──────────────────────────────────────────────
scikit-learn>=1.4.0
xgboost>=2.0.3
lightgbm>=4.3.0
transformers>=4.39.0
torch>=2.2.0
optuna>=3.6.0
mlflow>=2.11.0
joblib>=1.3.0

# ── Output / Infra ───────────────────────────────────────
plotly>=5.20.0
dash>=2.16.0
dash-bootstrap-components>=1.6.0
apscheduler>=3.10.4
celery>=5.3.6
sqlalchemy>=2.0.0
discord.py>=2.3.2
python-dotenv>=1.0.1

# ── Backtesting ──────────────────────────────────────────
vectorbt>=0.26.1

# ── Utilities ────────────────────────────────────────────
loguru>=0.7.2
pydantic>=2.6.0
tenacity>=8.2.3
# ── Testing ────────────────────────────────────────────
pytest
pytest-asyncio
asyncio
datetime
pathlib
zoneinfo